fix: int to long

+++ backend/Models/Place.cs
@@ -2,7 +2,7 @@ namespace backend.Models
 {
 public class Place
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Google_Id { get; set; } = string.Empty; 
         public string Name { get; set; } = string.Empty; 
         public decimal Lat { get; set; }
@@ -26,8 +26,8 @@ public class Place
         public string? Photo_7 { get; set; }
         public string? Photo_8 { get; set; }
         public string? Photo_9 { get; set; }
-        public int Num_Likes { get; set; } = 0;
-        public int Num_Dislikes { get; set; } = 0;
+        public long Num_Likes { get; set; } = 0;
+        public long Num_Dislikes { get; set; } = 0;
     }
 }
